Work in progress on refactoring unit tests
