+++ src/Console.Tests/Commands/CodeGenerationAssemblyCommandTests.cs
@@ -5,7 +5,7 @@ public class CodeGenerationAssemblyCommandTests
     public class Constructor
     {
         [Fact]
-        public void Throws_On_Null_Argument()
+        public void Throws_On_Null_Arguments()
         {
             typeof(CodeGenerationAssemblyCommand).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments();
         }
