+++ src/Console.Tests/Commands/CommandBaseTests.cs
@@ -24,7 +24,7 @@ public class CommandBaseTests
     public class Constructor
     {
         [Fact]
-        public void Throws_On_Null_Argument()
+        public void Throws_On_Null_Arguments()
         {
             typeof(CommandBaseTest).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments();
         }
