+++ src/Console.Tests/Commands/RunTemplateCommandTests.cs
@@ -5,7 +5,7 @@ public class RunTemplateCommandTests
     public class Constructor
     {
         [Fact]
-        public void Throws_On_Null_Argument()
+        public void Throws_On_Null_Arguments()
         {
             typeof(RunTemplateCommand).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments();
         }
