+++ src/Core.CodeGeneration.Tests/CodeGenerationAssemblySettingsTests.cs
@@ -5,7 +5,7 @@ public class CodeGenerationAssemblySettingsTests
     public class Constructor
     {
         [Fact]
-        public void Throws_On_Null_Argument()
+        public void Throws_On_Null_Arguments()
         {
             typeof(CodeGenerationAssemblySettings).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments(
                 parameterPredicate: p => !new[] { "currentDirectory", "classNameFilter" }.Contains(p.Name), // this parameter is optional
