+++ src/Core.CodeGeneration.Tests/CodeGenerationAssemblyTests.cs
@@ -5,7 +5,7 @@ public class CodeGenerationAssemblyTests
     public class Constructor
     {
         [Fact]
-        public void Throws_On_Null_Argument()
+        public void Throws_On_Null_Arguments()
         {
             typeof(CodeGenerationAssembly).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments();
         }
