+++ src/Core.CodeGeneration.Tests/CodeGenerationEngineTests.cs
@@ -5,7 +5,7 @@ public class CodeGenerationEngineTests
     public class Constructor
     {
         [Fact]
-        public void Throws_On_Null_Argument()
+        public void Throws_On_Null_Arguments()
         {
             typeof(CodeGenerationEngine).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments();
         }
