+++ src/Core.Tests/GenerationEnvironments/StringBuilderEnvironment/StringBuilderEnvironmentTests.Constructor.cs
@@ -5,7 +5,7 @@ public partial class StringBuilderEnvironmentTests
     public class Constructor : StringBuilderEnvironmentTests
     {
         [Fact]
-        public void Throws_On_Null_Argument()
+        public void Throws_On_Null_Arguments()
         {
             typeof(StringBuilderEnvironment).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments();
         }
