+++ src/Core.Tests/TemplateEngine/TemplateEngineTests.Constructor.cs
@@ -5,7 +5,7 @@ public partial class TemplateEngineTests
     public class Constructor : TemplateEngineTests
     {
         [Fact]
-        public void Throws_On_Null_Argument()
+        public void Throws_On_Null_Arguments()
         {
             typeof(TemplateEngine).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments();
         }
