+++ src/Core.Tests/TemplateEngineContextTests.cs
@@ -10,7 +10,7 @@ public class TemplateEngineContextTests
     public class Constructor : TemplateEngineContextTests
     {
         [Fact]
-        public void Throws_On_Null_Argument()
+        public void Throws_On_Null_Arguments()
         {
             typeof(TemplateEngineContext).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments();
         }
