+++ src/Core.Tests/TemplateIdentifiers/TemplateInstanceIdentifierTests.cs
@@ -5,7 +5,7 @@ public class TemplateInstanceIdentifierTests
     public class Constructor
     {
         [Fact]
-        public void Throws_On_Null_Argument()
+        public void Throws_On_Null_Arguments()
         {
             typeof(TemplateInstanceIdentifier).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments();
         }
