+++ src/Core.Tests/TemplateInitializerComponents/ModelInitializerTests.cs
@@ -13,7 +13,7 @@ public class ModelInitializerTests
     public class Constructor
     {
         [Fact]
-        public void Throws_On_Null_Argument()
+        public void Throws_On_Null_Arguments()
         {
             typeof(ModelInitializerComponent).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments();
         }
