+++ src/Core.Tests/TemplateInitializerComponents/ProviderPluginInitializerComponentTests.cs
@@ -12,7 +12,7 @@ public class ProviderPluginInitializerComponentTests
     public class Constructor
     {
         [Fact]
-        public void Throws_On_Null_Argument()
+        public void Throws_On_Null_Arguments()
         {
             typeof(ProviderPluginInitializerComponent).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments();
         }
