+++ src/Core.Tests/TemplateParameter/TemplateParameterTests.Constructor.cs
@@ -5,7 +5,7 @@ public class TemplateParameterTests
     public class Constructor
     {
         [Fact]
-        public void Throws_On_Null_Argument()
+        public void Throws_On_Null_Arguments()
         {
             typeof(TemplateParameter).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments();
         }
