+++ src/Core.Tests/TemplateParameterExtractorTests.cs
@@ -9,7 +9,7 @@ public class TemplateParameterExtractorTests
     public class Constructor
     {
         [Fact]
-        public void Throws_On_Null_Argument()
+        public void Throws_On_Null_Arguments()
         {
             typeof(TemplateParameterExtractor).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments();
         }
