+++ src/Core.Tests/TemplateProviderTests.cs
@@ -9,7 +9,7 @@ public class TemplateProviderTests
     public class Constructor
     {
         [Fact]
-        public void Throws_On_Null_Argument()
+        public void Throws_On_Null_Arguments()
         {
             typeof(TemplateProvider).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments();
         }
