+++ src/Core.Tests/TemplateRenderers/MultipleContentTemplateRendererTests.Constructor.cs
@@ -5,7 +5,7 @@ public partial class MultipleContentTemplateRendererTests
     public class Constructor : MultipleContentTemplateRendererTests
     {
         [Fact]
-        public void Throws_On_Null_Argument()
+        public void Throws_On_Null_Arguments()
         {
             typeof(MultipleContentTemplateRenderer).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments();
         }
