+++ src/Core.Tests/TemplateRenderers/StringBuilderTemplateRendererTests.Constructor.cs
@@ -5,7 +5,7 @@ public partial class StringBuilderTemplateRendererTests
     public class Constructor
     {
         [Fact]
-        public void Throws_On_Null_Argument()
+        public void Throws_On_Null_Arguments()
         {
             typeof(StringBuilderTemplateRenderer).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments();
         }
