+++ src/Core.Tests/ValueConverterTests.cs
@@ -9,7 +9,7 @@ public class ValueConverterTests
     public class Constructor
     {
         [Fact]
-        public void Throws_On_Null_Argument()
+        public void Throws_On_Null_Arguments()
         {
             typeof(ValueConverter).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments();
         }
