--- src/TemplateProviders.ChildTemplateProvider.Tests/Extensions/ServiceCollectionExtensions/ServiceCollectionExtensionsTests.AddTemplateFrameworkChildTemplateProvider.cs
@@ -1,21 +1,0 @@
-namespace TemplateFramework.TemplateProviders.ChildTemplateProvider.Tests.Extensions;
-
-public class ServiceCollectionExtensionsTests
-{
-    public class AddTemplateFrameworkChildTemplateProvider
-    {
-        [Fact]
-        public void All_Dependencies_Can_Be_Resolved()
-        {
-            // Act
-            using var provider = new ServiceCollection()
-                .AddTemplateFramework()
-                .AddTemplateFrameworkChildTemplateProvider()
-                .AddSingleton(Substitute.For<ITemplateComponentRegistryPluginFactory>())
-                .BuildServiceProvider(new ServiceProviderOptions { ValidateOnBuild = true, ValidateScopes = true });
-
-            // Assert
-            provider.Should().NotBeNull();
-        }
-    }
-}
