--- src/TemplateProviders.ChildTemplateProvider.Tests/ProviderComponent/ProviderComponentTests.cs
@@ -1,7 +1,0 @@
-﻿namespace TemplateFramework.TemplateProviders.ChildTemplateProvider.Tests;
-
-public partial class ProviderComponentTests
-{
-    protected ProviderComponent CreateSut() => new(new[] { TemplateCreatorMock });
-    protected ITemplateCreator TemplateCreatorMock { get; } = Substitute.For<ITemplateCreator>();
-}
