+++ src/TemplateProviders.ChildTemplateProvider.Tests/TemplateIdentifiers/TemplateByNameIdentifierTests.cs
@@ -3,7 +3,7 @@
 public class TemplateByNameIdentifierTests
 {
     [Fact]
-    public void Throws_On_Null_Argument()
+    public void Throws_On_Null_Arguments()
     {
         typeof(TemplateByNameIdentifier).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments();
     }
