--- src/TemplateProviders.CompiledTemplateProvider.Tests/ProviderComponent/ProviderComponentTests.cs
@@ -1,9 +1,0 @@
-﻿namespace TemplateFramework.TemplateProviders.CompiledTemplateProvider.Tests;
-
-public partial class ProviderComponentTests
-{
-    protected IAssemblyService AssemblyServiceMock { get; } = Substitute.For<IAssemblyService>();
-    protected ITemplateFactory CompiledTemplateFactoryMock { get; } = Substitute.For<ITemplateFactory>();
-
-    protected ProviderComponent CreateSut() => new(AssemblyServiceMock, CompiledTemplateFactoryMock);
-}
