+++ src/TemplateProviders.CompiledTemplateProvider.Tests/TemplateIdentifiers/CompiledTemplateTemplateIdentifierTests.cs
