+++ src/TemplateProviders.StringTemplateProvider.Tests/TemplateIdentifiers/ExpressionStringTemplateIdentifierTests.cs
@@ -5,7 +5,7 @@ public class ExpressionStringTemplateIdentifierTests
     public class Constructor
     {
         [Fact]
-        public void Throws_On_Null_Argument()
+        public void Throws_On_Null_Arguments()
         {
             typeof(ExpressionStringTemplateIdentifier).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments(p => !new[] { "pluginAssemblyName", "pluginClassName", "currentDirectory" }.Contains(p.Name));
         }
