+++ src/TemplateProviders.StringTemplateProvider.Tests/TemplateIdentifiers/FormattableStringTemplateIdentifierTests.cs
@@ -5,7 +5,7 @@ public class FormattableStringTemplateIdentifierTests
     public class Constructor
     {
         [Fact]
-        public void Throws_On_Null_Argument()
+        public void Throws_On_Null_Arguments()
         {
             typeof(FormattableStringTemplateIdentifier).ShouldThrowArgumentNullExceptionsInConstructorsOnNullArguments(p => !new[] { "pluginAssemblyName", "pluginClassName", "currentDirectory" }.Contains(p.Name));
         }
